set 3

+++ set3/IsInt.java
@@ -5,10 +5,10 @@ import java.util.*;
 public class IsInt{
     public static void main(String args[]){
         Scanner sc=new Scanner(System.in);
-        int n;
+        double n;
         String s=sc.next();
         try {
-            n=Integer.parseInt(s);
+            n=Double.parseDouble(s);
             System.out.println("yes");
         }
         catch(Exception e)
